backend/requirements.txt
Flask==2.2.2
requests==2.28.1
psycopg2-binary==2.9.3